+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/cache/configuration/UpstreamJavaConfigurationCacheKitTest.groovy
@@ -1,0 +1,59 @@
+package ru.vyarus.gradle.plugin.animalsniffer.cache.configuration
+
+import org.gradle.testkit.runner.BuildResult
+import org.gradle.testkit.runner.TaskOutcome
+import ru.vyarus.gradle.plugin.animalsniffer.AbstractKitTest
+import ru.vyarus.gradle.plugin.animalsniffer.UpstreamKitTest
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 12.12.2024
+ */
+class UpstreamJavaConfigurationCacheKitTest extends AbstractKitTest {
+
+    def "Check configuration cache support"() {
+
+        setup:
+        build """
+            plugins {
+                id 'java'
+                id 'ru.vyarus.animalsniffer'
+            }
+            
+            animalsniffer {
+                ignoreFailures = true                
+            }
+
+            repositories { mavenCentral()}
+            dependencies {
+                signature 'org.codehaus.mojo.signature:java16-sun:1.0@signature'
+                implementation 'org.slf4j:slf4j-api:1.7.25'
+            }
+        """
+
+        fileFromClasspath('src/main/java/invalid/Sample.java', '/ru/vyarus/gradle/plugin/animalsniffer/java/invalid/Sample.java')
+        //debug()
+
+        when: "run task"
+        BuildResult result = runVer(UpstreamKitTest.GRADLE_VERSION, 'check', '--configuration-cache', '--configuration-cache-problems=warn')
+
+        then: "no configuration cache incompatibilities"
+        result.output.contains("1 problem was found storing the configuration cache")
+        result.output.contains('Gradle runtime: support for using a Java agent with TestKit')
+        result.output.contains('Calculating task graph as no cached configuration is available for tasks:')
+
+        then: "task successful"
+        result.task(':animalsnifferMain').outcome == TaskOutcome.SUCCESS
+
+        when: "run from cache"
+        println '\n\n------------------- FROM CACHE ----------------------------------------'
+        result = runVer(UpstreamKitTest.GRADLE_VERSION, 'check', '--configuration-cache', '--configuration-cache-problems=warn')
+
+        then: "cache used"
+        result.output.contains('Reusing configuration cache.')
+
+//        then: "task successful"
+//        result.task(':animalsnifferMain').outcome == TaskOutcome.SUCCESS
+
+    }
+}