+++ CHANGELOG.md
@@ -18,7 +18,8 @@
 * Add new properties to AnimalSniffer task to simplify detailed configuration:
   - targetType: type of task (SourceSet, MultiplatformTarget, AndroidVariant)
   - targetName: name of target (used in animalsniffer task name)
-* Call animalsniffer directly instead of ant tasks 
+* Call animalsniffer directly instead of ant tasks
+* Configuration cache compatibility
 
 ### 1.7.2 (2024-11-18)
 * Update animalsniffer 1.23 -> 1.24